base_database class structure
